+++ packages/satyrographos.0.0.1.7/CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Haskell PVP](https://pvp.haskell.org/).
 
 ## [Unreleased]
 
+## [v0.0.1.7] - 2019-04-19
+### Fixed
+- Satyrographos does not fail even when OPAM does not exist
+
+## [v0.0.1.6] - 2019-04-09
+### Added
+- Add `install -copy` to copy files rather than create symlinks.
+
 ## [v0.0.1.5] - 2019-02-13
 ### Changed
 - When satysfi dist does not exist in the OPAM registory, Satyrographos use one in either `/usr/local/share/satysfi/dist` or `/usr/share/satysfi/dist`.
@@ -46,7 +54,9 @@ and this project adheres to [Haskell PVP](https://pvp.haskell.org/).
 - Detect duplicated package files
 - Detect duplicated hash definitions
 
-[Unreleased]: https://github.com/na4zagin3/satyrographos/compare/v0.0.1.5...HEAD
+[Unreleased]: https://github.com/na4zagin3/satyrographos/compare/v0.0.1.7...HEAD
+[v0.0.1.7]: https://github.com/na4zagin3/satyrographos/compare/v0.0.1.6...v0.0.1.7
+[v0.0.1.6]: https://github.com/na4zagin3/satyrographos/compare/v0.0.1.5...v0.0.1.6
 [v0.0.1.5]: https://github.com/na4zagin3/satyrographos/compare/v0.0.1.4...v0.0.1.5
 [v0.0.1.4]: https://github.com/na4zagin3/satyrographos/compare/v0.0.1.3...v0.0.1.4
 [v0.0.1.3]: https://github.com/na4zagin3/satyrographos/compare/v0.0.1.2...v0.0.1.3